ENH: Renamed SoftwareGuide.tex to remove ambiguitiy.
The name of the .tex file and the name of Directories
in the source an build tree were all the same, and this
made it difficult to use grep to find the desired items.

Additionally the LaTeX tools take the base name rather
than the full file name, and this caused more confusion while
debugging the programs builds.