ENH: Improve Hello World directions.
Add a better description of how to build and run the hello world example.

Suggest an out-of-source build.  Make a distinction between running CMake and
building with the generated files.

Make sure to indicate that the resulting executable should be run from the
command line.

Remove MSDEV reference (no longer common).

Change-Id: I12d003c3038d82bf9dbbc7240928eac07405a922