ENH: Revise Configure and Build Instructions
Minor edits improving the flow and consistency of the chapter.

Inserted an updated screenshot for ccmake interface.

Inserted au updated screenshot for cmake-gui interface on Windows.

Removed referecnes to MFC-based GUI on Windows.

Corrected the name of CMake GUI to cmake-gui.

Corrected the references to the names of buttons in CMake GUI.

In response to reviewers' comments: Revision of titles for this
chapter and some of its sections.

In response to reviewers' comments: Two small new subsections:
'Compiling ITK' and 'Installing ITK on Your System'.

The edits for the 'Advanced Module Configuration' section are
based on this Kitware Blog Post:
http://www.kitware.com/blog/home/post/557

In the 'Hello World' section the content of the CMakeLists.txt
is pasted in the .tex file. The content of the HelloWorld.cxx is
processed and the generated HelloWorld.tex is included. In this
case (unlike with all examples in the rest of the text) this
method doesn't seem to fit here, because the generated .tex file
has this line:

'The source code for this section can be found in the file
HelloWorld.cxx',

which does not fit the text properly, because here we are not at
the start of the section and the statement is somewhat incorrect
as it omits the mention of the CMakeList.txt file. Maybe in this
case it is worth pasting the code part of the HelloWorld.cxx
in the text here, just like it was done with CMakeLists.txt?

In addition the LaTeX comments in HelloWorld.cxx need a couple
minor corrections, like incorrect use of apostrophes and such.

As a general comment, there seems to be a lack of consistency with
naming different platforms in the text. For example, in many cases
the use of 'UNIX' implies UNIX and Linux. In some cases 'Linux' is
used to imply both. Perhaps some convention needs to be
established for this.

Change-Id: I9e056d3dd9163acbe309782ef937c6e6d56dacc5