ENH: Part one of converting Registration chapter to ITKv4
This patch includes the changes that are needed to the
registration chapter (chapter 3 of book 2) to reflect the ITKv4
registration framework.
Changes are completed before section 3.7 (Multi-Resolution
    Registration).
Next changes will be pushed in separate patches.
I broked down the conversion process to several patches, so it can be
reviewed more easily.

Change-Id: Ic4fb0a44d48fd6c6d1c376f28d3721f82bc0103b