ENH: Upsample when building with PDF_QUALITY_LEVEL "Printer".
The print service requires a minimal DPI for all images.  We bluff the DPI by
upsampling the images.

Imagemagick has difficulty with the metadata of some input images, like
BrainT1SliceBorder20.png, so specify the rescaling overriding the supposed
density of the input and output images, and specifying the scale with a pixel
limit.

Change-Id: I7cbbb2e371ce7d967f4d835779ec0a90d71391d8