ENH: Add option to build HTML output with tex4ht.
A new CMake configuration variable, GENERATE_HTML, which defaults to OFF, is
added.  When ON, HTML output will be generated for both books with the tex4ht
package.

Build was tested on Linux. A work-around was required to patch tex4ht when
using 20090611_p1038-r2 on Gentoo -- a link is added in the comments on how to
get the appendix to build.

Output is added to an 'html' directory.  A new 'index.html' file is created
that directs the reader to the generated content in Book1/ and Book2/.

Change-Id: I52408e546789155aad9bfd55a8c0aaa4fbe20ad3