BUG: Missing binary Examples folder
`RunExamples` has been moved from `ITKSoftwareGuide/SoftwareGuide/Examples`
to a CMake `ExternalProject_Add`. This means that the subfolder
`ITKSoftwareGuide/SoftwareGuide/Examples` is not created a configure time
anymore. When Examples are parsed to be converted to `tex` files, the parser
fails to write the generated `tex` file as the output folder does not
exist.