BUG: Fix compiler list time schedule format.
The compiler list time schedule in chapter `Configuring and Building
ITK` was being displayed with an awkward layout.

See for example p.10 in a recent build belonging to PR #64:
https://open.cdash.org/upload/8fc8cf811a130b76f66c2b81a3f3027ac17e3b09/ITKSoftwareGuide-Book1.pdf

As explained in:
https://en.wikibooks.org/wiki/LaTeX/Mathematics
the `\[` is a LaTeX shorthand for a displayed equation environment.

This topic removes the unnecessary backward slash escape character.