ENH: Make the section name appear in footnote 2 .
Footnote 2 on page 4 of the current ITK Software Guide PDF version
available at:
https://itk.org/ItkSoftwareGuide.pdf

refers to `Section 2.1. Obtaining the Software`. However, the footnote
only displays the section number `2.1`.

This patch set makes LaTeX display the section name, and also adds
additional information to make it clear that it refers to a resource
within the file itself.