DOC: Update copyright assignment to NumFOCUS
As part of the process of the Insight Software Consortium joining the
NumFOCUS non-profit corporation as a Fiscally Sponsored Project, all
assets, including intellectual property assets, are being transferred to
enable dissolution of the Insight Software Consortium non-profit
corporate legal entity.

The Insight Software Consortium Board of Directors resolved in the
2019-10-22 Meeting to assign ITK copyright to NumFOCUS and CMake and
IGSTK to Kitware. A copy of the signed copyright transfer documents can be
found in the corresponding pull request to ITK:
https://github.com/InsightSoftwareConsortium/ITK/pull/1598