DOC: Print backslash mark
Print backslash mark.

Fixes:
```
Note that ITK prefers the backslash ("
") style versus the at-sign (“@”) style to write the documentation command
```

and
```
When documenting a class header, if citations are required, use the Doxygen
`par References` command to list the references in a separate and
clearly visible paragraph
```

Use the `code` command for the at-sign as well for the sake of
consistency.