DOC: Fix PDF file properties in Book 2
Fix the PDF file properties: set the `SupressPDFMark` variable to `true`
in the `eps` files.

The properties of the eps figures generated by gnuplot at the time the
PDF is generated were ovewriting the LaTeX `hyperref` settings of the
book, effectively modifying the title, author, subject, and keywords of
the PDF file.

Relevant thread:
https://groups.google.com/g/comp.graphics.apps.gnuplot/c/WTjwj-lbC90